Add inputSelectGroup and generic inputWithType
inputSelectGroup is just like inputSelect but when you specify grouped
choices in the digestive-functors side, it creates optgroups in the
resulting input.

inputWithType is something that I ended up using to support some of
the more fancy HTML5 input types without having to write a specialized
input for each one, and you can see I actually rewrote a couple of the
existing inputs in terms of this more generic version. A good example
of this is the user can easily create an email input, telephone number
input, range input, etc without having to necessarily extend the
library itself.